Consolidate direct action unit tests into test_action.py
Move 14 tests that call droll.action functions directly (not through
player.apply or Game methods) into test_action.py where they belong:

- TestDefeatDragonHeroesInterchangeable (3 tests) from test_dragon.py
- TestDefeatDragonHeroesWildcard (3 tests) from test_dragon.py
- TestRerollParty (5 tests) from test_game.py
- TestRegroupDiscard (3 tests, new class) from test_world.py

Clean up now-unused imports in the source files.  Coverage of
droll/action.py is identical before and after (92%, same missing lines).
