+++ tests/test_dragon.py
@@ -7,7 +7,7 @@ from dataclasses import fields, replace
 import random
 import unittest
 
-from droll import action, error, heroes, player, struct, world
+from droll import error, heroes, player, struct, world
 
 
 class TestDragon(unittest.TestCase):
@@ -153,129 +153,4 @@ class TestDragon(unittest.TestCase):
                 "dragon",
                 "mage",
                 "mage",
-            )
-
-
-class TestDefeatDragonHeroesInterchangeable(unittest.TestCase):
-
-    def test_less_interesting_successful_cases(self):
-        """Test valid dragon defeats with interchangeable heroes."""
-        self.assertTrue(
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric", "thief", "mage", _interchangeable={"fighter"}
-            )
-        )
-        self.assertTrue(
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric", "thief", "fighter", _interchangeable={"fighter"}
-            )
-        )
-
-    def test_less_interesting_failure_cases(self):
-        """Test invalid dragon defeats with interchangeable heroes."""
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric", "thief", _interchangeable={"fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric", "fighter", _interchangeable={"fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric",
-                "thief",
-                "champion",
-                "mage",
-                _interchangeable={"fighter"},
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "cleric",
-                "thief",
-                "champion",
-                "fighter",
-                _interchangeable={"fighter"},
-            )
-
-    def test_more_interesting_failure_cases(self):
-        """Invalid dragon defeats with multiple interchangeable hero types."""
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "mage", "mage", "mage", _interchangeable={"mage", "fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "fighter",
-                "fighter",
-                "fighter",
-                _interchangeable={"mage", "fighter"},
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "fighter", "mage", "mage", _interchangeable={"mage", "fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_interchangeable(
-                "mage",
-                "fighter",
-                "fighter",
-                _interchangeable={"mage", "fighter"},
-            )
-
-
-class TestDefeatDragonHeroesWildcard(unittest.TestCase):
-
-    def test_less_interesting_successful_cases(self):
-        """Test valid dragon defeats with wildcard heroes."""
-        self.assertTrue(
-            action.defeat_dragon_heroes_wildcard(
-                "cleric",
-                "thief",
-                "mage",
-            )
-        )
-        self.assertTrue(
-            action.defeat_dragon_heroes_wildcard(
-                "cleric", "thief", "fighter", _wildcard={"scroll"}
-            )
-        )
-
-    def test_less_interesting_failure_cases(self):
-        """Test invalid dragon defeats with wildcard heroes."""
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "cleric", "thief", _wildcard={"scroll"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "cleric", "fighter", _wildcard={"fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "cleric", "thief", "champion", "mage", _wildcard={"fighter"}
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "cleric",
-                "thief",
-                "champion",
-                "fighter",
-                _wildcard={"fighter"},
-            )
-
-    def test_more_interesting_failure_cases(self):
-        """Test invalid dragon defeats with multiple wildcard heroes."""
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "mage",
-                "mage",
-                "scroll",
-            )
-        with self.assertRaises(error.DrollError):
-            action.defeat_dragon_heroes_wildcard(
-                "fighter",
-                "fighter",
-                "fighter",
-                _wildcard={"mage"},
             )
