+++ tests/test_game.py
@@ -7,7 +7,7 @@ from dataclasses import replace
 import random
 import unittest
 
-from droll import action, dice, struct, world
+from droll import struct
 from droll.error import DrollError
 from droll.game import Game, GameState
 from droll.player import Default
@@ -121,84 +121,3 @@ class TestGame(unittest.TestCase):
         g._world = replace(g._world, dungeon=struct.Dungeon())
         result = g.retire()
         self.assertEqual(result, GameState.STOP)
-
-
-class TestRerollParty(unittest.TestCase):
-    """Tests for action.reroll supporting party dice (issue #82)."""
-
-    def setUp(self):
-        """Fixture with a world containing known dungeon and party dice."""
-        self.world = replace(
-            world.new_world(),
-            dungeon=struct.Dungeon(goblin=2, skeleton=1),
-            party=struct.Party(fighter=2, cleric=1, scroll=2),
-        )
-
-    def _canned_randrange(self, sequence):
-        """Return a randrange function that yields predetermined values."""
-        seq = list(sequence)
-
-        def randrange(start, stop):
-            return start + seq.pop(0)
-
-        return randrange
-
-    def test_reroll_single_party_die(self):
-        """Rerolling a single party die removes it and re-rolls one party die."""
-        # Roll value 0 => fighter
-        randrange = self._canned_randrange([0])
-        result = action.reroll(
-            self.world, randrange, "scroll", "fighter"
-        )
-        # One scroll consumed for the hero cost
-        self.assertEqual(result.party.scroll, 1)
-        # Fighter was removed (2-1=1) then re-rolled as fighter (+1=2)
-        self.assertEqual(result.party.fighter, 2)
-        # Dungeon unchanged
-        self.assertEqual(result.dungeon, self.world.dungeon)
-
-    def test_reroll_multiple_party_dice(self):
-        """Rerolling multiple party dice removes and re-rolls each."""
-        # Roll values: 0 => fighter, 1 => cleric
-        randrange = self._canned_randrange([0, 1])
-        result = action.reroll(
-            self.world, randrange, "scroll", "fighter", "cleric"
-        )
-        # One scroll consumed
-        self.assertEqual(result.party.scroll, 1)
-        # fighter: 2 - 1 + 1(rolled) = 2; cleric: 1 - 1 + 1(rolled) = 1
-        self.assertEqual(result.party.fighter, 2)
-        self.assertEqual(result.party.cleric, 1)
-        # Dungeon unchanged
-        self.assertEqual(result.dungeon, self.world.dungeon)
-
-    def test_reroll_mix_dungeon_and_party(self):
-        """Rerolling a mix of dungeon and party dice handles both."""
-        # Dungeon roll: value 0 => goblin; Party roll: value 2 => mage
-        randrange = self._canned_randrange([0, 2])
-        result = action.reroll(
-            self.world, randrange, "scroll", "goblin", "fighter"
-        )
-        # One scroll consumed
-        self.assertEqual(result.party.scroll, 1)
-        # Dungeon: goblin was 2, removed 1, re-rolled as goblin (+1) => 2
-        self.assertEqual(result.dungeon.goblin, 2)
-        # Party: fighter was 2, removed 1, re-rolled as mage (+1)
-        self.assertEqual(result.party.fighter, 1)
-        self.assertEqual(result.party.mage, 1)
-
-    def test_reroll_party_die_not_present_raises(self):
-        """Rerolling a party die that is not present raises DrollError."""
-        randrange = self._canned_randrange([])
-        with self.assertRaises(DrollError):
-            action.reroll(
-                self.world, randrange, "scroll", "mage"
-            )
-
-    def test_reroll_unknown_target_raises(self):
-        """Rerolling a target that is neither dungeon nor party raises DrollError."""
-        randrange = self._canned_randrange([])
-        with self.assertRaises(DrollError):
-            action.reroll(
-                self.world, randrange, "scroll", "nonexistent"
-            )
